requirements.txt
soccerdata==1.8.7
numpy
pandas
pyyaml
requests
python-dateutil
scipy

